Jelambrar96 cloud composer

+++ gfunctions/download_function/main.py
@@ -9,6 +9,15 @@ import requests
 from google.cloud.storage import Client
 
 
+def bucket_exists(bucket_name):
+    storage_client = Client()
+    try:
+        bucket = storage_client.get_bucket(bucket_name)
+        return True
+    except:
+        return False
+
+
 def download_gh_archive_data(request):
 
     request_json = request.get_json(silent=True)
@@ -19,30 +28,42 @@ def download_gh_archive_data(request):
     elif request_args and 'date' in request_args:
         date = request_args['date']
     else:
-        yesterday_datetime = datetime.strptime(date, "%Y-%m-%d") - timedelta(days=1)
+        yesterday_datetime = datetime.now() - timedelta(days=1)
         yesterday_date = yesterday_datetime.strftime("%Y-%m-%d")
         date = yesterday_date
 
+
+    if request_json and 'hour' in request_json:
+        hour = request_json['hour']
+    elif request_args and 'hour' in request_args:
+        hour = request_args['hour']
+    else:
+        hour = datetime.now().hour
+        
+
     storage_client = Client()
     bucket_name = os.environ.get('BUCKET_NAME')
+
+    flag_bucket_exist = bucket_exists(bucket_name)
+    if not flag_bucket_exist:
+        return f"ERROR: Bucket {bucket_name} NOT FOUND", 404
+
     bucket = storage_client.bucket(bucket_name)
 
     with tempfile.TemporaryDirectory() as output_dir:
 
-        for hour in range(24):
+        url = f"https://data.gharchive.org/{date}-{hour}.json.gz"
+        file_name = f"{date}-{hour}.json.gz"
+        file_path = os.path.join(output_dir, file_name)
+        response = requests.get(url)
+        with open(file_path, 'wb') as f:
+            f.write(response.content)
 
-            url = f"https://data.gharchive.org/{date}-{hour}.json.gz"
-            file_name = f"{date}-{hour}.json.gz"
-            file_path = os.path.join(output_dir, file_name)
-            response = requests.get(url)
-            with open(file_path, 'wb') as f:
-                f.write(response.content)
+        blob = bucket.blob(f"gh-archives/raw/{date}/")
+        blob.upload_from_string('')
 
-            blob = bucket.blob(f"gh-archives/raw/{date}/")
-            blob.upload_from_string('')
-
-            blob_name = f"gh-archives/raw/{date}/{file_name}"
-            blob = bucket.blob(blob_name)
-            blob.upload_from_filename(file_path)
+        blob_name = f"gh-archives/raw/{date}/{file_name}"
+        blob = bucket.blob(blob_name)
+        blob.upload_from_filename(file_path)
 
     return "Success: download_gh_archive_data. ", 200
